requirements.txt
openai==1.12.0
python-dotenv==1.0.0
pandas==2.1.4
plotly==5.24.1